fixing small compilation issue on getting nb of phys pages in mac

+++ cobs/util/misc.cpp
@@ -23,7 +23,7 @@ uint64_t get_phys_pages () {
   if (phys_pages == 0) {
 #ifdef __APPLE__
     uint64_t mem;
-    uint64_t len = sizeof(mem);
+    size_t len = sizeof(mem);
     sysctlbyname("hw.memsize", &mem, &len, NULL, 0);
     phys_pages = mem/sysconf(_SC_PAGE_SIZE);
 #endif
